requirements-production.txt
# GMAO Sistema - Dependencias de Producción
# Para deployment en Google Cloud - Disfood

# Framework principal
Flask==2.3.3
Werkzeug==2.3.7

# Base de datos
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7  # PostgreSQL driver para Cloud SQL

# Autenticación y seguridad
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Limiter==3.5.0

# Google Cloud Platform
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-sql-connector==1.5.0

# Cache y almacenamiento
redis==4.6.0
Flask-Caching==2.1.0

# Servidor WSGI para producción
gunicorn==21.2.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.0.1  # Para procesamiento de imágenes
openpyxl==3.1.2  # Para exportar Excel
reportlab==4.0.4  # Para generar PDFs

# Monitoreo y logging
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0

# Desarrollo y testing (solo para CI/CD)
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Seguridad adicional
cryptography==41.0.7
bcrypt==4.0.1